Change processor to use config file and output json
Add parser for a config file where we can specify the adjustable values. This change also adds a config file with sections for simple diff, diff with pixel threshold, and histogram diff. We also add logic after the processing to store all the results for one video in a json.