Add edge diff to list of processors
Adds logic to compare the diffs of only the edges of each frame and returns whether the fraction of changed pixels is greater than a threshold. Threshold and minVal/maxVal of intensities (to decide whether a pixel is part of an edge) are retrieved from config file. This also edits diff_settings.config to try a couple values for intensities (detailed vs. rough edges) and thresholds