Add min motion area to list of processors
For each frame, we now want to look at the areas that changed when we use motion detection to compare to the last frame sent. This change adds a processor that gets the contour changes (as rectangles) from the last sent frame and checks the size. If the size of any of the rectangles is above a certain threshold, send the frame. We also add a few options for the threshold in the config file (i.e. are we looking for cars (big box) or pedestrians (smaller box)?)