Change motion area processor to use fraction of total pixels
We were just using a raw number of pixels for the threshold of the min area with motion to decide whether to send the frame. It should be compared to the total to account for different frame sizes. Edit the logic in MinMotionArea to do this and edit the config file to use fractions instead